requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
PyGithub
langchain
langchain-google-genai
